refactor: move middlewares test to a test folder
