+++ src/main/middlewares/__tests__/bodyParser.test.ts
@@ -1,6 +1,6 @@
 import request from 'supertest'
 
-import app from '../config/app'
+import app from '../../config/app'
 
 describe('Body Parser middleware', () => {
   it('Should parse body as json', async () => {
