+++ src/main/middlewares/__tests__/cors.test.ts
@@ -1,6 +1,6 @@
 import request from 'supertest'
 
-import app from '../config/app'
+import app from '../../config/app'
 
 describe('CORS middleware', () => {
   it('Should enable cors', async () => {
